Rename lrpar's `lex` module to `lex_api`.
As much as I'm indifferent to the term "API", it at least makes clearer
that lrpar can't do any lexing itself, and that lrpar is merely defining
the terms on which external lexers can interface to lrpar.
